Handle object hierarchy (closes #16).
Set child and parent CityObjects as attributes. In FME we don't have/set
an explicit object hierarchy, but each feature is on the same level.
Therefore we store the child-parent relationships in attributes,
for those who are interested in the hierarchies. Just like in cityjson.
I'm not adding the children and parents attributes to the schema,
because its better if they are hidden from the
table view, since there can be many-many children for each feature.