Remove SemanticSurface hierarchy (closes #17)
Having a hierarchy of semantic surfaces (child-parent) is not really
necessary. I didn't find data that actually uses this concept, because
it is mainly relevant for LoD3 or higher.
This commit logs a Warning in case a SemanticSurface has a children or
parent attribute and discards these attributes from the data.
Discussed with @hugoledoux.